feat: Add user update functionality
Introduces a new `updateUser` function to the DataContext. This function allows for partial updates to user details, including sanitizing name and email inputs and automatically generating an avatar URL if the name changes.

Also refactors the `RequestList` page to remove assignee selection during request creation, as this functionality was not fully implemented or used.

Finally, it prepares the `AdminUsers` page for future user editing by adding state and UI elements for an edit modal.